[IMP] Added support for forced updates.
By setting one of the force update modes (ForceUpdateOnly, ForceRunOnly, ForceReadyState), a "Force Ready" state is written to the curq_state_history table. If the table does not exist yet, it's created. This way it's possible to deal with failed updates (which don't write Ready when done) and adopting this system to existing databases which don't have the table yet, and thus don't have Ready state, preventing updates.